Fix several problems with the data-array
['message'] has to be lowercase to allow for substituion within logstash.
That field should be parsed as json and the payload-field "message" is
then automagically replaced by the contents of the message-key of the
jsonified array.

['HttpHost'] Changed to always contain string types. The original one could
contain booleans as the function returns string|boolean and that lead
to problems with mixed derived datatypes in elasticsearch.

Problems with high log_level entries (>= Zend_Log::DEBUG) if Cm_Redis_Session
was used. Those log levels caused every call to the session Model to generate
a log entry that contains a call to logging. En Voila an infinite loop was
created. We only use the RedisSession Methods if the log_level is low enough